refactor (core): add a high-level API to manager layers
This commits changes the way layers are added to the scene; the goal is to build
a clearer API by being explicit instead of generic (addColorLayer vs addLayer).

With this change, Providers enrich the layers they preprocess with new methods
allowing the user to know which layer types can be added.
For now, only TileProvider does such a thing, but in the future we could
imagine adding a 'addColorProvider' to colorize a pointcloud or 'addElevationProvider'
to position 3dtiles elements.