feat(core): add vector tile loading in TMS
Vector tiles can be loaded in TMS as textures. Only Mapbox Vector Tile
is supported for now. This commit adds a new Parser (VectorTileParser),
and a new Helper (similar to OGCWebServiceHelper) VectorTileHelper.

Modifications have been done in Feature2Texture to support style as a
function, and as an object.

Also add two examples showing the use of Mapbox Vector Tiles in iTowns.
The first one is loading a full vector tile layer in a Planar view,
while the second only display national parks on top of a simple Ortho
layer in a a Globe view.